// src/types/Credits.ts
export interface Person {
  adult: boolean;
  gender: number;
  id: number;
  known_for_department: string;
  name: string;
  original_name: string;
  popularity: number;
  profile_path: string | null;
  cast_id?: number;
  character?: string;
  credit_id: string;
  department?: string;
  job?: string;
}
export interface CreditsProps {
  id: number;
  cast: Person[];
  crew: Person[];
}